functions/requirements.txt
azure-functions
requests
pymongo
azure.storage
re